slugify_component moved to utils
